[STEP19] test: 상품 목록 조회 컨트롤러 작성 및 테스트

+++ src/main/java/kr/hhplus/be/server/domain/product/controller/dto/TopProductRanking.java
@@ -1,7 +1,9 @@
 package kr.hhplus.be.server.domain.product.controller.dto;
 
+import java.math.BigDecimal;
+
 public record TopProductRanking(
         Long productId,
         String productName,
-        double productPrice,
+        BigDecimal productPrice,
         long soldCount) {}
